Refactor sql route draft
