+++ features/shadow/core/src/main/java/org/apache/shardingsphere/shadow/route/ShadowSQLRouter.java
@@ -20,7 +20,8 @@ package org.apache.shardingsphere.shadow.route;
 import org.apache.shardingsphere.infra.annotation.HighFrequencyInvocation;
 import org.apache.shardingsphere.infra.config.props.ConfigurationProperties;
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
-import org.apache.shardingsphere.infra.route.DecorateSQLRouter;
+import org.apache.shardingsphere.infra.route.type.DataSourceSQLRouter;
+import org.apache.shardingsphere.infra.route.type.DecorateSQLRouter;
 import org.apache.shardingsphere.infra.route.context.RouteContext;
 import org.apache.shardingsphere.infra.route.context.RouteMapper;
 import org.apache.shardingsphere.infra.route.context.RouteUnit;
@@ -38,7 +39,7 @@ import java.util.Optional;
  * Shadow SQL router.
  */
 @HighFrequencyInvocation
-public final class ShadowSQLRouter implements DecorateSQLRouter<ShadowRule> {
+public final class ShadowSQLRouter implements DecorateSQLRouter<ShadowRule>, DataSourceSQLRouter<ShadowRule> {
     
     @Override
     public void decorateRouteContext(final RouteContext routeContext, final QueryContext queryContext, final ShardingSphereDatabase database,
