+++ features/sharding/core/src/main/java/org/apache/shardingsphere/sharding/route/engine/ShardingSQLRouter.java
@@ -22,7 +22,8 @@ import org.apache.shardingsphere.infra.binder.context.type.CursorAvailable;
 import org.apache.shardingsphere.infra.config.props.ConfigurationProperties;
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
 import org.apache.shardingsphere.infra.metadata.database.rule.RuleMetaData;
-import org.apache.shardingsphere.infra.route.EntranceSQLRouter;
+import org.apache.shardingsphere.infra.route.type.EntranceSQLRouter;
+import org.apache.shardingsphere.infra.route.type.TableSQLRouter;
 import org.apache.shardingsphere.infra.route.context.RouteContext;
 import org.apache.shardingsphere.infra.session.query.QueryContext;
 import org.apache.shardingsphere.sharding.cache.route.CachedShardingSQLRouter;
@@ -45,7 +46,7 @@ import java.util.Optional;
  * Sharding SQL router.
  */
 @HighFrequencyInvocation
-public final class ShardingSQLRouter implements EntranceSQLRouter<ShardingRule> {
+public final class ShardingSQLRouter implements EntranceSQLRouter<ShardingRule>, TableSQLRouter<ShardingRule> {
     
     @Override
     public RouteContext createRouteContext(final QueryContext queryContext, final RuleMetaData globalRuleMetaData, final ShardingSphereDatabase database,
