+++ infra/route/src/main/java/org/apache/shardingsphere/infra/route/type/EntranceSQLRouter.java
@@ -15,29 +15,33 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.route.engine.type;
+package org.apache.shardingsphere.infra.route.type;
 
+import org.apache.shardingsphere.infra.config.props.ConfigurationProperties;
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
 import org.apache.shardingsphere.infra.metadata.database.rule.RuleMetaData;
+import org.apache.shardingsphere.infra.route.SQLRouter;
 import org.apache.shardingsphere.infra.route.context.RouteContext;
-import org.apache.shardingsphere.infra.route.context.RouteMapper;
-import org.apache.shardingsphere.infra.route.context.RouteUnit;
-import org.apache.shardingsphere.infra.route.engine.SQLRouteExecutor;
+import org.apache.shardingsphere.infra.rule.ShardingSphereRule;
 import org.apache.shardingsphere.infra.session.query.QueryContext;
 
-import java.util.Collections;
-
 /**
- * All SQL route executor.
+ * Entrance SQL Router.
+ * 
+ * @param <T> type of rule
  */
-public final class AllSQLRouteExecutor implements SQLRouteExecutor {
+public interface EntranceSQLRouter<T extends ShardingSphereRule> extends SQLRouter<T> {
     
-    @Override
-    public RouteContext route(final QueryContext queryContext, final RuleMetaData globalRuleMetaData, final ShardingSphereDatabase database) {
-        RouteContext result = new RouteContext();
-        for (String each : database.getResourceMetaData().getStorageUnits().keySet()) {
-            result.getRouteUnits().add(new RouteUnit(new RouteMapper(each, each), Collections.emptyList()));
-        }
-        return result;
-    }
+    /**
+     * Create route context.
+     *
+     * @param queryContext query context
+     * @param globalRuleMetaData global rule meta data
+     * @param database database
+     * @param rule rule
+     * @param props configuration properties
+     * @return route context
+     */
+    RouteContext createRouteContext(QueryContext queryContext,
+                                    RuleMetaData globalRuleMetaData, ShardingSphereDatabase database, T rule, ConfigurationProperties props);
 }
