--- infra/route/src/test/java/org/apache/shardingsphere/infra/route/engine/dialect/MySQLSQLRouteExecutorDeciderTest.java
@@ -1,38 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.infra.route.engine.dialect;
-
-import org.apache.shardingsphere.sql.parser.statement.core.statement.SQLStatement;
-import org.apache.shardingsphere.sql.parser.statement.mysql.dal.MySQLShowTableStatusStatement;
-import org.apache.shardingsphere.sql.parser.statement.mysql.dal.MySQLShowTablesStatement;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.mock;
-
-class MySQLSQLRouteExecutorDeciderTest {
-    
-    @Test
-    void assertIsNeedRouteAll() {
-        MySQLSQLRouteExecutorDecider decider = new MySQLSQLRouteExecutorDecider();
-        assertTrue(decider.isNeedRouteAll(new MySQLShowTablesStatement()));
-        assertTrue(decider.isNeedRouteAll(new MySQLShowTableStatusStatement()));
-        assertFalse(decider.isNeedRouteAll(mock(SQLStatement.class)));
-    }
-}
